The last commit before depoy

+++ app/Http/Controllers/ExportController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Models\Export;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
+use Inertia\Inertia;
 
 class ExportController extends Controller
 {
@@ -14,12 +16,16 @@ class ExportController extends Controller
      */
     public function index()
     {
-        $exports = Export::paginate(15);
-
-
+        $exports = Export::paginate(5);
+        return Inertia::render('Reports', [
+            'exports' => $exports
+        ]);
     }
+    public function show($export) {
+        $export = Export::find($export);
 
-
+        return Storage::download($export->file_name);
+    }
 
     /**
      * Remove the specified resource from storage.
@@ -27,11 +33,16 @@ class ExportController extends Controller
      * @param  \App\Models\Export  $export
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Export $export)
+    public function destroy($export)
     {
-        Storage::delete($export->file_name);
-        $export->delete();
+        $export = Export::find($export);
 
-        return 'Deletado';
+        if ($export) {
+            Storage::delete($export->file_name);
+            $export->delete();
+        }
+
+        return redirect()->back()
+            ->with('success', 'Seu arquivo foi removido com sucesso');
     }
 }
